fixed flaky test
